Refactor the way protocols are found.
The new architecture is based on a polymorphic class hierarchy.  This
allows every aspect of the protocol-finding process to be customized in
any number of different ways.  In particular, I wanted to make it easy
to get protocols from various locations on the internet in the future.

The new class hierarchy can also be pickled and passed through pipes,
so that not every process in a pipeline needs to search the whole
filesystem for protocols.  This didn't turn out to be a significant
optimization, but it won't hurt.

Fixes #19, #17, #12