Load python protocols without starting a new process.
Most stepwise protocols are assembled from a number smaller
sub-protocols, and most of these sub-protocols are scripts written in
python.  Prior to this commit, stepwise would launch a new python
process every time it encountered such a sub-protocol.  This turned out
to be a significant bottleneck because python can take several hundered
milliseconds to start up.  This is especially true when heavy packages
like `pkg_resources` and `pandas` are being used, which many stepwise
protocols do.

This commit improves performance by executing protocols that are written
in python (identified by a '.py' file extension) directly in the current
python process, so that the start-up cost is only paid once.