Improve error checking for Quantity.
The motivation for this was that I had a hard time figuring out what
went wrong when I accidentally provided a Quantity as the "value" of a
quantity.  The error wasn't detected right away, so the code didn't fail
until some arithmetic operation in the middle of rendering the reagent
table.  To address this, I added more stringent type checking to the
Quantity constructor.