Flush before and after redirecting stdout.
This is less of a problem when stdout is attached to a TTY, because then
it's flushed after every newline.  But when stdout is attached to a
pipe, it isn't flushed as frequently and output that is written while
stdout is being redirected may not be flushed until afterwards.

Note also that the buffering in question is happening in python.  Pipes
themselves are unbuffered, but higher-level file abstractions
(`sys.stdout` in python, `*FILE` in C) do their own buffering to
minimize the number of system calls they need to make.