feat: don't always require stock concentrations to be quantities
Stock concentrations still have to be quantities (i.e. numeric values
with units) if they will be used to scale a reagent up or down.  But if
they won't be used in that way, they are now allowed to be any string,
which may be more expressive in some cases.