fix: truncate all lines when printing
Overly-long lines can mess up the whole document by affecting where the
pages break.  Therefore, I've decided to truncate any lines that would
extend past the page, not just such lines in the header.  While this
does mean that important information can be truncated, that information
is still available in the text-file version of the protocol.