feat: add API for creating steps and substeps
Protocol steps can now be either strings or objects.  Objects can
provide methods to control certain aspects of their formatting, namely
wrapping.  Strings are used verbatim.

This is meant to make it easier to create properly-wrapped steps and
substeps in python protocols.  Previously, it wasn't really possible to
do this correctly.  You'd have to wrap the steps in the code, but
dynamic changes like variable substitutions or step numbering could
easily lead to lines that were too long.  The step API wraps the text
once the correct width is known, solving this problem.