fix: check the header before attempting to unpickle a byte stream
The pickle module can't reliably tell whether or not a byte stream is
actually a pickle, because although the binary protocols (>=1) have a
defined header, the text protocol (0) does not.  To avoid this problem,
we have to check for the header ourselves, which we can do because we
know we're using a binary protocol.