feat: add better text formatting tools
These tools are:

- paragraph_list
- unordered_list
- ordered_list
- preformatted

A wide variety of formats can be achieved by piecing together these
primitives.  Additionally, strings are now treated as paragraphs rather
than as verbatim text.  This means that strings are wrapped by default,
making it unnecessary to use any formatting tools at all in the most
common cases.  In cases where text really shouldn't be wrapped (e.g.
read from a text file), the `preformatted()` class should be used.

Fixes #35

This commit also fixes an issue with `sw note` that caused it to fail on
formatted text, and moved the ability to insert footnotes in the manner
of that command to the `Protocol` class (to be more easily accessible
from python).

Fixes #41