feat: don't namespace plugin config files
I realized that it would be potentially useful to create a plugin that
does nothing but provide new configs for another plugin.  For example, a
vendor might distribute stepwise_mol_bio settings for their products.
This would've been impossible prior to this commit, because the config
files for each plugin were forced into unique namespaces.

This commit also:

- Adds tests for the StepwiseConfig and PresetConfig classes.
- Adds support for the new Config.with_options() class method.