fix: remove duplicate footnotes
Fixes #56

This commit is a bit of a hack, because any API user who might create
duplicate footnotes will need to manually call the new
`deduplicate_footnotes()` to get rid of them.  A more robust solution
would take care of that automatically, but I feel like that's a bigger
project.