fix: favor shorter names when matching tags.
Consider matching the names 'b' and 'a/b' with the tag 'b'.  Previously,
both would match equally well an the user would be required to be more
specific.  But there's no way to be more specific about the first name;
it's already a perfect match!  So I modified the tag-matching algorithm
to prefer shorter names, since longer names can always be specified more
completely.