feat: support multi-step master mixes
This commit add the `Reactions` class, which (i) calculates how to break
any reaction into an arbitrary number of master mixes and (ii) formats a
protocol describing how to setup the reaction.  This class will
ultimately replace the `MasterMix` class, but for now I'm leaving both
in place to allow for an easier transition.  Fixes #59.