feat: make reactions intrinsically ordered
Previously, the data structure underlying the `Reaction` class was a
dictionary.  This had the downside that new reagents could only be
added to the end of the reaction.  It was also difficult to change a
reagent's key with changing its order.  This commit changes that data
structure to a list, which avoids these downsides and fits better with
the way we think of reactions.

Changes to the `Reaction` class:

- Add `append`, `prepend`, `insert`, and `remove` methods.

- Make the solvent less of a special case.  The solvent now starts as a
  normal reagent and is converted into a solvent by one of a few means,
  e.g. setting the volume "to" the total reaction volume.

- Allow reagents to be copied between and within reactions more
  readily, e.g. via `__setitem__`.

- Remove the `order` attribute, which is no longer needed.

Changes to the `Reactions` class:

- Avoid creating master mixes with only a single reagent.

- Allow making extra of even the final reaction in a series of master
  mixes.

- Use more consistent rule for naming master mixes.

Fixes #55, #60