feat: put flagged reagents in earlier master mixes
The purpose is to allow the user to identify important or
difficult-to-pipet reagents, e.g. enzymes in glycerol, and to instruct
the master mix algorithm to include them in earlier master mixes where
there will be larger volumes and more consistency.

This commit also refactors how mixes are scored, to make adding new
score terms more straight-forward.