feat: override keys in reagent tables
Previously, keys were specified via the "Reagent" column of reaction
tables.  But I realized that I use this column more like a "Name"
column, since I specify long, clear names that are meant to be displayed
in the protocol.  For protocols that require specific keys for specific
reagents, it makes more sense to be able to specify those in their own
column.