fix: consider adding multiple reagents to each master mix
Previously, the master mix algorithm would only consider adding one
reagent to each mix (except the first).  In some cases, this prevents
the best solution from being found.

The new algorithm solves this problem by introducing the concept of a
"partial mix": a collection of reagents that will be added at the same
time, but that may also be expanded later.

This commit also changes how the algorithm decides which reagents to
merge on each step.  Previously, all pairs of reagents in the top two
levels (but not the same level, unless matching) would be merged.  Now,
all pairs that have the smallest number of joint combinations are
merged.  This expresses what I want to do better, and I think it results
in many fewer combinations ultimately being sampled.